Major improvement to ICs for observed histories. Support data subsegments
Before: For observes hists, initial conditions (ICs, values before t0) 
were never changed from their initialization.
Now: The first few data points are used to set ICs. The inferred trace 
is shortened accordingly, and the model only integrated as far as the 
data allow.

This commit also includes fixes to allow fitting on a data segment 
shorter than the model attached to the optimizer. This was motivated by 
the need to drop data bins to set ICs. As a bonus, we can now fit using 
data segments of different lengths.