More flexible specification of latent hists in default workflow.
The default workflow assumes are latent hists are part of the 'input' 
submodel. This commit allows them to be specified using other submodels, 
as long as they link to a history of 'input'.
For example, if 'input.ξ' is connected to 'dynamics.I', then the 
arguments to the workflow can specify either 'input.ξ' or 'dynamics.I' 
as a latent history.
This was the documented behaviour, but the actual behaviour was 
previously to accept only 'input.ξ'.