Use `accumulate` instead of `static_accumulate` for latent logp
`static_accumulate` prevents any update function from being included in
the graph, which is not what we want in general. I think my original
reasoning was to prevent gradients from backpropagating through the
latent history, but this can be accomplished by locking it. Leaving the
dynamics unlocked allows gradients to propagate through them, which we
want.

Also, to prevent the latent fit from converging on a flat latent, the
Optimize_WF_template removes the latent objective when fitting latents
(the latent objective *is* included when fitting the parameters)

Finally, `ObjectiveFunction + 0` just returns the objective function
itself. This allows to use `sum()` without creating nested classes when
not necessary.